Refactor RestClient package structure, reformat code for consistency, and clean up unused imports across modules
