+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/data/network/KeypleSyncEndPointClient.kt
@@ -13,7 +13,6 @@
 package org.calypsonet.keyple.demo.reload.remote.data.network
 
 import io.reactivex.Single
-import org.calypsonet.keyple.demo.reload.remote.domain.spi.RestClient
 import org.eclipse.keyple.distributed.MessageDto
 import org.eclipse.keyple.distributed.spi.SyncEndpointClientSpi
