+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/ui/activities/AbstractDemoActivity.kt
@@ -18,9 +18,9 @@ import dagger.android.support.DaggerAppCompatActivity
 import javax.inject.Inject
 import org.calypsonet.keyple.demo.reload.remote.R
 import org.calypsonet.keyple.demo.reload.remote.data.SharedPrefDataRepository
+import org.calypsonet.keyple.demo.reload.remote.data.network.RestClient
 import org.calypsonet.keyple.demo.reload.remote.databinding.ToolbarBinding
 import org.calypsonet.keyple.demo.reload.remote.domain.model.ServerStatusEvent
-import org.calypsonet.keyple.demo.reload.remote.domain.spi.RestClient
 import org.calypsonet.keyple.demo.reload.remote.ui.model.UiSamStatus
 import org.eclipse.keyple.core.util.json.JsonUtil
 import org.greenrobot.eventbus.EventBus
